# Conflicts:
#	compare/src/main/java/com/technoserv/rest/resources/SkudResource.java

+++ common/src/main/java/com/technoserv/db/dao/objectmodel/api/RequestDao.java
@@ -21,5 +21,5 @@ public interface RequestDao extends Dao<Long, Request> {
 
     List<Request> findByCriteria(RequestSearchCriteria criteria);
 
-    Integer countByCriteria(RequestSearchCriteria criteria);
+    Long countByCriteria(RequestSearchCriteria criteria);
 }
